chore: fixup dangerjs rule
DSYS-3660
