+++ .danger/__fixtures__/changeset/lost-cradle-unlimited.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/avatar': major
+'@twilio-paste/core': minor
+---
+
+Test changeset with minor core and major other
